swap to client commands
