--- Assets/AirshipPackages/@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModKickCommand.ts.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ab0c65f162a8246fe9c478d2aa4b050c
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: c4bd260abc8845c88b54ec021581f2a0, type: 3}
-  icon: {instanceID: 0}
