--- Assets/AirshipPackages/@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModMuteCommand.ts.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c3df53600bbee4162bff37dd5822f0cd
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: c4bd260abc8845c88b54ec021581f2a0, type: 3}
-  icon: {instanceID: 0}
