--- Assets/AirshipPackages/@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModUnbanCommand.ts
@@ -1,48 +1,0 @@
-import { Airship } from "@Easy/Core/Shared/Airship";
-import { ChatCommand } from "@Easy/Core/Shared/Commands/ChatCommand";
-import { Game } from "@Easy/Core/Shared/Game";
-import { Player } from "@Easy/Core/Shared/Player/Player";
-import { ChatColor } from "@Easy/Core/Shared/Util/ChatColor";
-import { ModerationServiceDatabaseTypes } from "@Easy/Core/Shared/TypePackages/moderation-service-types";
-import { ModerationServiceBridgeTopics, ServerBridgeApiGameModRemoveAction } from "@Easy/Core/Server/ProtectedServices/Airship/Moderation/ModerationService";
-import { GameModerationCommand } from "./GameModCommandHelper";
-
-export class GameModUnbanCommand extends ChatCommand {
-    constructor() {
-        super(GameModerationCommand.UNBAN, [], "/unban <username> <reason?>", "Unbans a player.", true);
-        this.requiresPermission = true; // Granted by having moderation role and remove ban permission for current game
-    }
-
-    public Execute(modPlayer: Player, args: string[]): void {
-        if (args.size() === 0) {
-            modPlayer.SendMessage(ChatColor.Red(`Invalid usage: /unban <username> <reason?>`));
-            return;
-        }
-
-        const targetUsername = args[0];
-        const reason = args[1];
-
-        const target = Airship.Players.FindByFuzzySearch(targetUsername);
-        if (!target) {
-            modPlayer.SendMessage(ChatColor.Red(`Player not found: ${targetUsername}`));
-            return;
-        }
-        
-        const action = contextbridge.invoke<ServerBridgeApiGameModRemoveAction>(
-            ModerationServiceBridgeTopics.GameModerationRemoveAction,
-            LuauContext.Protected,
-            modPlayer.userId,
-            {
-                actionType: ModerationServiceDatabaseTypes.GameModerationActionType.BAN,
-                uid: target.userId,
-                gameId: Game.gameId,
-                reason
-            }
-        );
-        if (action) {
-            modPlayer.SendMessage(`Unbanned ${target.username} for reason: ${action.reason}`);
-        } else {
-            modPlayer.SendMessage(ChatColor.Red(`Failed to unban ${target.username}.`));
-        }
-    }
-}
