--- Assets/AirshipPackages/@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModUnbanCommand.ts.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ee402c46db0ad4528a6bc1e2a4d43179
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: c4bd260abc8845c88b54ec021581f2a0, type: 3}
-  icon: {instanceID: 0}
