--- Assets/AirshipPackages/@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModUnmuteCommand.ts.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a6abb252550c54eb9b79d0eb3fef6fe0
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: c4bd260abc8845c88b54ec021581f2a0, type: 3}
-  icon: {instanceID: 0}
