+++ Assets/AirshipPackages/@Easy/Core/Shared/Chat/AirshipChatSingleton.ts
@@ -30,15 +30,6 @@ import { ChatColor } from "../Util/ChatColor";
 import { ChatUtil } from "../Util/ChatUtil";
 import ObjectUtils from "../Util/ObjectUtils";
 import { Signal } from "../Util/Signal";
-import { GameModBanCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModBanCommand";
-import { GameModKickCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModKickCommand";
-import { GameModMuteCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModMuteCommand";
-import { GameModUnbanCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModUnbanCommand";
-import { GameModUnmuteCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModUnmuteCommand";
-import { GameModLookupCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModLookupCommand";
-import { GameModNoteCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModNoteCommand";
-import { GameModTempBanCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModTempBanCommand";
-import { GameModTempMuteCommand } from "@Easy/Core/Server/Services/Chat/Commands/GameMod/GameModTempMuteCommand";
 
 class ChatMessageEvent extends Cancellable {
 	/**
@@ -225,17 +216,6 @@ export class AirshipChatSingleton {
 		this.RegisterCommand(new HelpCommand());
 		this.RegisterCommand(new KickCommand());
 		this.RegisterCommand(new TeamsCommand());
-
-		/** Mod Commands */
-		this.RegisterCommand(new GameModLookupCommand());
-		this.RegisterCommand(new GameModKickCommand());
-		this.RegisterCommand(new GameModMuteCommand());
-		this.RegisterCommand(new GameModBanCommand());
-		this.RegisterCommand(new GameModTempMuteCommand());
-		this.RegisterCommand(new GameModTempBanCommand());
-		this.RegisterCommand(new GameModUnmuteCommand());
-		this.RegisterCommand(new GameModUnbanCommand());
-		this.RegisterCommand(new GameModNoteCommand());
 
 	}
